[DomainDrivenDesign] Move coupled location to rest of Rector in standalone package
