+++ composer.json
@@ -40,6 +40,7 @@
             "Rector\\ContributorTools\\": "packages/ContributorTools/src",
             "Rector\\ConsoleDiffer\\": "packages/ConsoleDiffer/src",
             "Rector\\CodeQuality\\": "packages/CodeQuality/src",
+            "Rector\\DomainDrivenDesign\\": "packages/DomainDrivenDesign/src",
             "Rector\\NodeTypeResolver\\": "packages/NodeTypeResolver/src",
             "Rector\\Symfony\\": "packages/Symfony/src",
             "Rector\\CakePHP\\": "packages/CakePHP/src",
@@ -62,6 +63,7 @@
             "Rector\\NodeTypeResolver\\Tests\\": "packages/NodeTypeResolver/tests",
             "Rector\\CakePHP\\Tests\\": "packages/CakePHP/tests",
             "Rector\\CodeQuality\\Tests\\": "packages/CodeQuality/tests",
+            "Rector\\DomainDrivenDesign\\Tests\\": "packages/DomainDrivenDesign/tests",
             "Rector\\Php\\Tests\\": "packages/Php/tests",
             "Rector\\Symfony\\Tests\\": "packages/Symfony/tests",
             "Rector\\Silverstripe\\Tests\\": "packages/Silverstripe/tests",
@@ -94,8 +96,8 @@
             "tests/Issues/Issue594/Wrong",
             "tests/Rector/Assign/PropertyAssignToMethodCallRector/Wrong",
             "tests/Rector/RectorOrder/Wrong",
-            "tests/Rector/DomainDrivenDesign/ValueObjectRemoverRector/Wrong",
-            "tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/Wrong",
+            "packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverRector/Wrong",
+            "packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverDocBlockRector/Wrong",
             "tests/Rector/StaticCall/StaticCallToFunctionRector/Wrong",
             "tests/Rector/MethodCall/MethodCallToAnotherMethodCallWithArgumentsRector/Wrong",
             "tests/Rector/MethodCall/StaticMethodNameReplacerRector/Wrong",
