+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverDocBlockRector/Source/SomeValueObject.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverDocBlockRector\Source;
+
+class SomeValueObject
+{
+
+}
