+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverDocBlockRector/ValueObjectRemoverDocBlockRectorTest.php
@@ -1,0 +1,32 @@
+<?php declare(strict_types=1);
+
+namespace Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverDocBlockRector;
+
+use Iterator;
+use Rector\Testing\PHPUnit\AbstractRectorTestCase;
+
+/**
+ * @covers \Rector\DomainDrivenDesign\Rector\ValueObjectRemover\ValueObjectRemoverDocBlockRector
+ */
+final class ValueObjectRemoverDocBlockRectorTest extends AbstractRectorTestCase
+{
+    /**
+     * @dataProvider provideWrongToFixedFiles()
+     */
+    public function test(string $wrong, string $fixed): void
+    {
+        $this->doTestFileMatchesExpectedContent($wrong, $fixed);
+    }
+
+    public function provideWrongToFixedFiles(): Iterator
+    {
+        yield [__DIR__ . '/Wrong/wrong.php', __DIR__ . '/Correct/correct.php'];
+        yield [__DIR__ . '/Wrong/wrong2.php', __DIR__ . '/Correct/correct2.php'];
+        yield [__DIR__ . '/Wrong/wrong3.php', __DIR__ . '/Correct/correct3.php'];
+    }
+
+    protected function provideConfig(): string
+    {
+        return __DIR__ . '/config.yml';
+    }
+}
