+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverDocBlockRector/Wrong/wrong.php
