+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverDocBlockRector/Wrong/wrong2.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace SomeNamespace;
+
+use Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverDocBlockRector\Source\SomeValueObject;
+
+class FourthActionClass
+{
+    public function someFunction(?SomeValueObject $name): ?SomeValueObject
+    {
+        /** @var SomeValueObject|null $someValueObject */
+        $someValueObject = new SomeValueObject('value');
+    }
+}
