+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverDocBlockRector/Wrong/wrong3.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace SomeNamespace;
+
+use Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverDocBlockRector\Source\SomeValueObject;
+
+class ThirdActionClass
+{
+    /**
+     * @param null|SomeValueObject $name
+     */
+    public function someFunction(?SomeValueObject $name): ?SomeValueObject
+    {
+    }
+}
