+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverDocBlockRector/config.yml
@@ -1,0 +1,3 @@
+services:
+    Rector\DomainDrivenDesign\Rector\ValueObjectRemover\ValueObjectRemoverDocBlockRector:
+        'Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverDocBlockRector\Source\SomeValueObject': 'string'
