+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverRector/Correct/correct2.php.inc
