+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverRector/Correct/correct3.php.inc
@@ -1,0 +1,15 @@
+<?php
+
+namespace SomeNamespace;
+
+use Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverRector\Source\SomeValueObject;
+
+class ThirdActionClass
+{
+    /**
+     * @param null|SomeValueObject $name
+     */
+    public function someFunction(?string $name): ?string
+    {
+    }
+}
