+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverRector/Correct/correct4.php.inc
@@ -1,0 +1,14 @@
+<?php
+
+namespace SomeNamespace;
+
+use Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverRector\Source\SomeValueObject;
+
+class FourthActionClass
+{
+    public function someFunction(?string $name): ?string
+    {
+        /** @var SomeValueObject|null $someValueObject */
+        $someValueObject = 'value';
+    }
+}
