+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverRector/Source/SomeChildOfValueObject.php
@@ -1,0 +1,7 @@
+<?php declare(strict_types=1);
+
+namespace Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverRector\Source;
+
+class SomeChildOfValueObject extends SomeValueObject
+{
+}
