+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverRector/Source/SomeValueObject.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverRector\Source;
+
+class SomeValueObject
+{
+
+}
