+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverRector/Wrong/wrong2.php.inc
@@ -1,0 +1,10 @@
+<?php
+
+namespace SomeNamespace;
+
+class SecondActionClass
+{
+    public function someFunction(\Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverRector\Source\SomeChildOfValueObject $name)
+    {
+    }
+}
