+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverRector/Wrong/wrong4.php.inc
@@ -1,0 +1,14 @@
+<?php
+
+namespace SomeNamespace;
+
+use Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverRector\Source\SomeValueObject;
+
+class FourthActionClass
+{
+    public function someFunction(?SomeValueObject $name): ?SomeValueObject
+    {
+        /** @var SomeValueObject|null $someValueObject */
+        $someValueObject = new SomeValueObject('value');
+    }
+}
