+++ packages/DomainDrivenDesign/tests/Rector/ValueObjectRemoverRector/config.yml
@@ -1,0 +1,3 @@
+services:
+    Rector\DomainDrivenDesign\Rector\ValueObjectRemover\ValueObjectRemoverRector:
+        'Rector\DomainDrivenDesign\Tests\Rector\ValueObjectRemoverRector\Source\SomeValueObject': 'string'
