--- tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/Correct/correct.php
@@ -1,20 +1,0 @@
-<?php
-
-namespace Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector;
-
-use Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector\Source\SomeChildOfValueObject;
-
-class ActionClass
-{
-    /**
-     * @var string|null
-     */
-    private $someChildValueObject;
-
-    public function someFunction()
-    {
-        $this->someChildValueObject = new SomeChildOfValueObject('value');
-
-        $someChildValueObject = new SomeChildOfValueObject();
-    }
-}
