--- tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/Wrong/wrong2.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace SomeNamespace;
-
-use Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector\Source\SomeValueObject;
-
-class FourthActionClass
-{
-    public function someFunction(?SomeValueObject $name): ?SomeValueObject
-    {
-        /** @var SomeValueObject|null $someValueObject */
-        $someValueObject = new SomeValueObject('value');
-    }
-}
