--- tests/Rector/DomainDrivenDesign/ValueObjectRemoverDocBlockRector/Wrong/wrong3.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace SomeNamespace;
-
-use Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverDocBlockRector\Source\SomeValueObject;
-
-class ThirdActionClass
-{
-    /**
-     * @param null|SomeValueObject $name
-     */
-    public function someFunction(?SomeValueObject $name): ?SomeValueObject
-    {
-    }
-}
