--- tests/Rector/DomainDrivenDesign/ValueObjectRemoverRector/Correct/correct3.php.inc
@@ -1,15 +1,0 @@
-<?php
-
-namespace SomeNamespace;
-
-use Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverRector\Source\SomeValueObject;
-
-class ThirdActionClass
-{
-    /**
-     * @param null|SomeValueObject $name
-     */
-    public function someFunction(?string $name): ?string
-    {
-    }
-}
