--- tests/Rector/DomainDrivenDesign/ValueObjectRemoverRector/Source/SomeChildOfValueObject.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverRector\Source;
-
-class SomeChildOfValueObject extends SomeValueObject
-{
-}
