--- tests/Rector/DomainDrivenDesign/ValueObjectRemoverRector/Source/SomeValueObject.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverRector\Source;
-
-class SomeValueObject
-{
-
-}
