--- tests/Rector/DomainDrivenDesign/ValueObjectRemoverRector/Wrong/wrong2.php.inc
@@ -1,10 +1,0 @@
-<?php
-
-namespace SomeNamespace;
-
-class SecondActionClass
-{
-    public function someFunction(\Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverRector\Source\SomeChildOfValueObject $name)
-    {
-    }
-}
