--- tests/Rector/DomainDrivenDesign/ValueObjectRemoverRector/Wrong/wrong4.php.inc
@@ -1,14 +1,0 @@
-<?php
-
-namespace SomeNamespace;
-
-use Rector\Tests\Rector\DomainDrivenDesign\ValueObjectRemoverRector\Source\SomeValueObject;
-
-class FourthActionClass
-{
-    public function someFunction(?SomeValueObject $name): ?SomeValueObject
-    {
-        /** @var SomeValueObject|null $someValueObject */
-        $someValueObject = new SomeValueObject('value');
-    }
-}
